+++ src/index.ts
@@ -1,3 +1,2 @@
-import {Application} from './app/application';
-
-Application.run();
+export { Application } from './app/application';
+export { CliApplication } from './index-cli';