+++ src/utils/defaults.ts
@@ -1,0 +1,9 @@
+export const COMPODOC_DEFAULTS = {
+    title: 'Application documentation',
+    additionalEntryName: 'Additional documentation',
+    additionalEntryPath: 'additional-documentation',
+    folder: './documentation/',
+    port: 8080,
+    theme: 'gitbook',
+    base: '/'
+}