+++ test/src/todomvc-ng2/app/home/home.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 
 /**
- * The about component
+ * The home component
  */
 @Component({
     selector: 'home',